+++ src/ca/gfx_cmdbuffer.h
@@ -27,8 +27,8 @@ namespace ca
 		void cmdbuffer_bind_pipeline(cmdbuffer_t * cmdbuffer, pipeline_t * pipeline);
 		void cmdbuffer_bind_indexbuffer(cmdbuffer_t * cmdbuffer, buffer_t * buffer, size_t offset, size_t stride);
 		void cmdbuffer_bind_vertexbuffer(cmdbuffer_t * cmdbuffer, buffer_t * buffer, size_t offset);
-		void cmdbuffer_set_viewport(cmdbuffer_t * cmdbuffer, u32 x, u32 y, u32 width, u32 height);
-		void cmdbuffer_set_scissor(cmdbuffer_t * cmdbuffer, u32 x, u32 y, u32 width, u32 height);
+		void cmdbuffer_set_viewport(cmdbuffer_t * cmdbuffer, f32 x, f32 y, f32 width, f32 height);
+		void cmdbuffer_set_scissor(cmdbuffer_t * cmdbuffer, i32 x, i32 y, u32 width, u32 height);
 		void cmdbuffer_draw(cmdbuffer_t * cmdbuffer, u32 vertex_start, u32 vertex_count);
 		void cmdbuffer_draw_indexed(cmdbuffer_t * cmdbuffer, u32 vertex_start, u32 index_start, u32 index_count);
 		void cmdbuffer_end_renderpass(cmdbuffer_t * cmdbuffer);
